TIP-1159: move product completeness model into the correct directory
