+++ src/Akeneo/Pim/Enrichment/Component/Product/Completeness/Model/ProductCompletenessCollection.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Component\Product\Model\Projection;
+namespace Akeneo\Pim\Enrichment\Component\Product\Completeness\Model;
 
 /**
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class ProductCompletenessWithMissingAttributeCodesCollection implements \IteratorAggregate, \Countable
+final class ProductCompletenessCollection implements \IteratorAggregate, \Countable
 {
     /** @var int */
     private $productId;
 
-    /** @var ProductCompletenessWithMissingAttributeCodes[] */
+    /** @var ProductCompleteness[] */
     private $completenesses = [];
 
     public function __construct(int $productId, array $completenesses)
@@ -34,26 +34,21 @@ final class ProductCompletenessWithMissingAttributeCodesCollection implements \I
         return new \ArrayIterator($this->completenesses);
     }
 
-    public function isEmpty(): bool
-    {
-        return count($this->completenesses) === 0;
-    }
-
-    public function getCompletenessForChannelAndLocale(string $channelCode, string $localeCode): ?ProductCompletenessWithMissingAttributeCodes
+    public function getCompletenessForChannelAndLocale(string $channelCode, string $localeCode): ?ProductCompleteness
     {
         $key = sprintf('%s-%s', $channelCode, $localeCode);
 
         return $this->completenesses[$key] ?? null;
     }
 
-    public function count()
-    {
-        return count($this->completenesses);
-    }
-
-    private function add(ProductCompletenessWithMissingAttributeCodes $completeness): void
+    private function add(ProductCompleteness $completeness): void
     {
         $key = sprintf('%s-%s', $completeness->channelCode(), $completeness->localeCode());
         $this->completenesses[$key] = $completeness;
+    }
+
+    public function count(): int
+    {
+        return count($this->completenesses);
     }
 }
