+++ src/Akeneo/Pim/Enrichment/Component/Product/Normalizer/Indexing/ProductCompletenessCollectionNormalizer.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Pim\Enrichment\Component\Product\Normalizer\Indexing;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessCollection;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessCollection;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
