+++ src/Akeneo/Pim/Enrichment/Component/Product/Normalizer/InternalApi/ProductCompletenessNormalizer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Enrichment\Component\Product\Normalizer\InternalApi;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessWithMissingAttributeCodes;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodes;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
