+++ src/Akeneo/Pim/Enrichment/Component/Product/Query/SaveProductCompletenesses.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Enrichment\Component\Product\Query;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessWithMissingAttributeCodesCollection;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodesCollection;
 
 /**
  * @author    Mathias METAYER <mathias.metayer@akeneo.com>
