+++ src/Akeneo/Pim/Structure/Component/Query/PublicApi/Family/RequiredAttributesMaskForChannelAndLocale.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Pim\Structure\Component\Query\PublicApi\Family;
 
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\CompletenessProductMask;
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessWithMissingAttributeCodes;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodes;
 
 /**
  * @author Pierre Allard <pierre.allard@akeneo.com>
