+++ tests/back/Pim/Enrichment/Integration/Completeness/AttributeType/IdentifierAttributeTypeCompletenessIntegration.php
@@ -3,7 +3,7 @@
 namespace AkeneoTest\Pim\Enrichment\Integration\Completeness\AttributeType;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessWithMissingAttributeCodes;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodes;
 use Akeneo\Pim\Structure\Component\Model\FamilyInterface;
 
 /**
