+++ tests/back/Pim/Enrichment/Integration/Completeness/CompletenessForLocalisableAttributeIntegration.php
@@ -3,7 +3,7 @@
 namespace AkeneoTest\Pim\Enrichment\Integration\Completeness;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompleteness;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompleteness;
 use Akeneo\Pim\Structure\Component\AttributeTypes;
 
 /**
