+++ tests/back/Pim/Structure/Specification/Component/Query/PublicApi/Family/RequiredAttributesMaskForChannelAndLocaleSpec.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Specification\Akeneo\Pim\Structure\Component\Query\PublicApi\Family;
 
 use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\CompletenessProductMask;
-use Akeneo\Pim\Enrichment\Component\Product\Model\Projection\ProductCompletenessWithMissingAttributeCodes;
+use Akeneo\Pim\Enrichment\Component\Product\Completeness\Model\ProductCompletenessWithMissingAttributeCodes;
 use PhpSpec\ObjectBehavior;
 
 final class RequiredAttributesMaskForChannelAndLocaleSpec extends ObjectBehavior
